% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram}
\alias{dendrogram}
\alias{as.dendrogram.hclust}
\alias{as.dendrogram.hdbscan}
\alias{reachability-coersion}
\alias{as.reachability}
\alias{as.dendrogram.reachability}
\title{Dendrogram Coersions}
\usage{
\method{as.dendrogram}{hclust}(object, ...)

\method{as.dendrogram}{hdbscan}(object, ...)

\method{as.dendrogram}{reachability}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{further arguments}
}
\description{
Coerce \link{dendrogram} objects to \link{hclust}, \link{hdbscan} and \link{reachability} objects.
}
\details{
Notes:
\itemize{
\item the coersion to hclust is a faster reimplementation of
\code{\link[stats:dendrogram]{stats::as.dendrogram()}}.
\item the coersion to hdbscan builds the non-simplified HDBSCAN hierarchy as a
dendrogram object.
}
}
