% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS3.R
\docType{data}
\name{DS3}
\alias{DS3}
\title{DS3: Spatial data with arbitrary shapes}
\format{
A data.frame with 8000 observations on the following 2 columns:
\describe{
\item{X}{a numeric vector}
\item{Y}{a numeric vector}
}
}
\source{
Obtained from the Clustering Datasets Repository, University of Joensuu, Finland.
}
\description{
Contains 8000 2-d points, with 6 "natural" looking shapes, all of which have
an sinusoid-like shape that intersects with each cluster.
The data set was originally used as a benchmark data set for the Chameleon clustering
algorithm (Karypis, Han and Kumar, 1999) to
illustrate the a data set containing arbitrarily shaped
spatial data surrounded by both noise and artifacts.
}
\examples{
data(DS3)
plot(DS3, pch = 20, cex = 0.25)
}
\references{
Karypis, George, Eui-Hong Han, and Vipin Kumar (1999).
Chameleon: Hierarchical clustering using dynamic modeling. \emph{Computer}
32(8): 68-75.
}
\keyword{datasets}
