% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcv.R
\name{dbcv}
\alias{dbcv}
\alias{DBCV}
\title{Density-Based Clustering Validation Index (DBCV)}
\usage{
dbcv(x, cl, d = 2)
}
\arguments{
\item{x}{a data matrix or a dist object.}

\item{cl}{a clustering (e.g., a integer vector)}

\item{d}{dimensionality of the original data if a dist object is provided.}
}
\value{
A list with the DBCV score for the clustering, the DSC values,
the DSPC distance matrix
and the cluster validity indices.
}
\description{
Calculate the Density-Based Clustering Validation Index (DBCV)  for a
clustering.
}
\details{
DBCV (Moulavi et al, 2014) computes a score based on the density sparseness of each cluster
and the density separation of each pair of clusters.

The density sparseness of a cluster (DSC) is deﬁned as the maximum edge weight of
a minimal spanning tree for the points of the cluster using the mutual
reachability distance. The
density separation of a pair of clusters (DSPC)
is deﬁned as the minimum reachability distance between the internal nodes of
the spanning trees of the two clusters.
The validity index for a cluster is calculated for these measures and aggregated
to a validity index for the whole clustering using a weighted average.

The index is in the range \eqn{[-1,1]}. If the cluster density compactness is better
than the density separation, a positive value is returned.

Noise points are included in the calculation only in the weighted average,
therefore clustering with more noise points will get a lower index.

\strong{Performance note:} This implementation calculates a distance matrix and thus
can only be used for small or sampled datasets.
}
\examples{
data(moons)

cl <- dbscan(moons, eps = .3, minPts = 3)
clplot(moons, cl)

dbcv(moons, cl)
}
\references{
Davoud Moulavi and Pablo A. Jaskowiak and
Ricardo J. G. B. Campello and Arthur Zimek and Jörg Sander (2014).
Density-Based Clustering Validation. In
\emph{Proceedings of the 2014 SIAM International Conference on Data Mining,}
pages 839-847
\doi{10.1137/1.9781611973440.96}
}
\author{
Matt Piekenbrock and Michael Hahsler
}
\concept{Evaluation Functions}
