% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram}
\alias{dendrogram}
\alias{as.dendrogram.hclust}
\alias{as.dendrogram.hdbscan}
\alias{reachability-coersion}
\alias{as.reachability}
\alias{as.dendrogram.reachability}
\title{Coersions to Dendrogram}
\usage{
\method{as.dendrogram}{hclust}(object, ...)

\method{as.dendrogram}{hdbscan}(object, ...)

as.reachability(object, ...)

\method{as.dendrogram}{reachability}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{further arguments}
}
\description{
Coerce \link{hclust}, \link{hdbscan} or \link{reachability} objects to \link{dendrogram}.
}
\details{
The coersion from hclust is a faster reimplementation of \code{\link[stats:dendrogram]{stats::as.dendrogram()}}.

The coersion from \link{hdbscan} builds the non-simplified HDBSCAN hierarchy as a
dendrogram object.
}
